fix(agent): repair OpenAI-gated unit tests for CI
Route LangChain-only tests through @mcp-use/agent/langchain with autoInitialize, move memory tests to integration, and exclude docs fixtures from test:unit.
